+++ android/src/main/java/com/idlefish/flutterboost/containers/FlutterContainerManager.java
@@ -8,6 +8,7 @@ import android.app.Activity;
 import android.os.Build;
 import android.util.Log;
 
+import com.idlefish.flutterboost.FlutterBoost;
 import com.idlefish.flutterboost.FlutterBoostUtils;
 
 import java.util.HashMap;
@@ -24,25 +25,33 @@ public class FlutterContainerManager {
 
     private static class LazyHolder {
         static final FlutterContainerManager INSTANCE = new FlutterContainerManager();
+        static {
+            INSTANCE.allContainers.put(FlutterBoost.ENGINE_ID, new HashMap<>());
+        }
     }
 
     public static FlutterContainerManager instance() {
         return FlutterContainerManager.LazyHolder.INSTANCE;
     }
 
-    private final Map<String, FlutterViewContainer> allContainers = new HashMap<>();
+    private final Map<String, Map<String, FlutterViewContainer>> allContainers = new HashMap<>();
     private final LinkedList<FlutterViewContainer> activeContainers = new LinkedList<>();
 
     // onContainerCreated
     public void addContainer(String uniqueId, FlutterViewContainer container) {
-        allContainers.put(uniqueId, container);
+        Map<String, FlutterViewContainer> containerMap = allContainers.putIfAbsent(container.getCachedEngineId(), new HashMap<>());
+        if(containerMap == null) {
+            containerMap = allContainers.get(container.getCachedEngineId());
+        }
+        containerMap.put(uniqueId, container);
         if (isDebugLoggingEnabled()) Log.d(TAG, "#addContainer: " + uniqueId + ", " + this);
     }
 
     // onContainerAppeared
     public void activateContainer(String uniqueId, FlutterViewContainer container) {
         if (uniqueId == null || container == null) return;
-        assert(allContainers.containsKey(uniqueId));
+        Map<String, FlutterViewContainer> containerMap = allContainers.get(FlutterBoost.ENGINE_ID);
+        assert(containerMap.containsKey(uniqueId));
 
         if (activeContainers.contains(container)) {
             activeContainers.remove(container);
@@ -54,17 +63,25 @@ public class FlutterContainerManager {
     // onContainerDestroyed
     public void removeContainer(String uniqueId) {
         if (uniqueId == null) return;
-        FlutterViewContainer container = allContainers.remove(uniqueId);
-        activeContainers.remove(container);
+        final FlutterViewContainer[] container = {null};
+        allContainers.forEach((engineId, containerMap) -> {
+            if (containerMap.containsKey(uniqueId)) {
+                container[0] = containerMap.remove(uniqueId);
+            }
+        });
+        activeContainers.remove(container[0]);
         if (isDebugLoggingEnabled()) Log.d(TAG, "#removeContainer: " + uniqueId + ", " + this);
     }
 
 
     public FlutterViewContainer findContainerById(String uniqueId) {
-        if (allContainers.containsKey(uniqueId)) {
-            return allContainers.get(uniqueId);
-        }
-        return null;
+        final FlutterViewContainer[] flutterViewContainer = {null};
+        allContainers.forEach((engineId, containerMap) -> {
+            if (containerMap.containsKey(uniqueId)) {
+                flutterViewContainer[0] = containerMap.get(uniqueId);
+            }
+        });
+       return flutterViewContainer[0];
     }
 
     public boolean isActiveContainer(FlutterViewContainer container) {
@@ -101,7 +118,11 @@ public class FlutterContainerManager {
     }
 
     public int getContainerSize() {
-        return allContainers.size();
+        return allContainers.get(FlutterBoost.ENGINE_ID).size();
+    }
+    public int getContainerSize(String engineId) {
+        Map<String, FlutterViewContainer> containerMap = allContainers.get(engineId);
+        return containerMap==null?0:containerMap.size();
     }
 
     @Override