+++ android/src/main/java/com/idlefish/flutterboost/containers/FlutterTextureHooker.java
@@ -12,16 +12,19 @@ import android.view.TextureView;
 import com.idlefish.flutterboost.FlutterBoost;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 
 import io.flutter.embedding.android.FlutterTextureView;
 import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.embedding.engine.FlutterEngineCache;
 import io.flutter.embedding.engine.renderer.FlutterRenderer;
 
 
 class FlutterTextureHooker {
-    private SurfaceTexture     restoreSurface;
+    private SurfaceTexture restoreSurface;
     private FlutterTextureView flutterTextureView;
-    private boolean            isNeedRestoreState = false;
+    private boolean isNeedRestoreState = false;
+    private String cachedEngineId;
 
     /**
      * Release surface when Activity.onDestroy / Fragment.onDestroy.
@@ -29,11 +32,13 @@ class FlutterTextureHooker {
      */
     public void onFlutterTextureViewRelease() {
         if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.M) {
-            int containerSize = FlutterContainerManager.instance().getContainerSize();
+            int containerSize = FlutterContainerManager.instance().getContainerSize(this.cachedEngineId);
             if (containerSize == 1) {
-                FlutterEngine   engine   = FlutterBoost.instance().getEngine();
-                FlutterRenderer renderer = engine.getRenderer();
-                renderer.stopRenderingToSurface();
+                FlutterEngine engine = FlutterEngineCache.getInstance().get(this.cachedEngineId);
+                if (engine != null) {
+                    FlutterRenderer renderer = engine.getRenderer();
+                    renderer.stopRenderingToSurface();
+                }
             }
             if (restoreSurface != null) {
                 restoreSurface.release();
@@ -50,19 +55,29 @@ class FlutterTextureHooker {
         if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.M) {
             if (restoreSurface != null && flutterTextureView != null && isNeedRestoreState) {
                 try {
-                    Class<? extends FlutterTextureView> aClass                         = flutterTextureView.getClass();
-                    Field                               isSurfaceAvailableForRendering = aClass.getDeclaredField("isSurfaceAvailableForRendering");
+                    Class<? extends FlutterTextureView> aClass = flutterTextureView.getClass();
+
+                    Field isSurfaceAvailableForRendering = aClass.getDeclaredField(
+                            "isSurfaceAvailableForRendering");
                     isSurfaceAvailableForRendering.setAccessible(true);
                     isSurfaceAvailableForRendering.set(flutterTextureView, true);
-
-                    Field isAttachedToFlutterRenderer = aClass.getDeclaredField("isAttachedToFlutterRenderer");
-                    isAttachedToFlutterRenderer.setAccessible(true);
-                    if (isAttachedToFlutterRenderer.getBoolean(flutterTextureView)) {
-                        FlutterEngine engine = FlutterBoost.instance().getEngine();
+                    boolean next = false;
+                    try {
+                        Field isAttachedToFlutterRenderer = aClass.getDeclaredField(
+                                "isAttachedToFlutterRenderer");
+                        isAttachedToFlutterRenderer.setAccessible(true);
+                        next = isAttachedToFlutterRenderer.getBoolean(flutterTextureView);
+                    } catch (NoSuchFieldException ignore) {
+                        Method shouldNotify = aClass.getDeclaredMethod("shouldNotify");
+                        shouldNotify.setAccessible(true);
+                        next = (Boolean) shouldNotify.invoke(flutterTextureView);
+                    }
+                    if (next) {
+                        FlutterEngine engine = FlutterEngineCache.getInstance().get(cachedEngineId);
                         if (engine != null) {
 
                             FlutterRenderer flutterRenderer = engine.getRenderer();
-                            Surface         surface         = new Surface(restoreSurface);
+                            Surface surface = new Surface(restoreSurface);
                             flutterRenderer.startRenderingToSurface(surface, false);
 
                             try {
@@ -85,9 +100,10 @@ class FlutterTextureHooker {
     /**
      * Hook FlutterTextureView for os version below Android.M.
      */
-    public void hookFlutterTextureView(FlutterTextureView flutterTextureView) {
+    public void hookFlutterTextureView(FlutterTextureView flutterTextureView, String cachedEngineId) {
+        this.cachedEngineId = cachedEngineId;
         if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.M) {
-            if(flutterTextureView!=null){
+            if (flutterTextureView != null) {
                 TextureView.SurfaceTextureListener surfaceTextureListener = flutterTextureView.getSurfaceTextureListener();
                 this.flutterTextureView = flutterTextureView;
                 this.flutterTextureView.setSurfaceTextureListener(new TextureView.SurfaceTextureListener() {
@@ -105,13 +121,12 @@ class FlutterTextureHooker {
                     @Override
                     public boolean onSurfaceTextureDestroyed(SurfaceTexture surface) {
                         try {
-                            Class<? extends FlutterTextureView> aClass                         = flutterTextureView.getClass();
-                            Field                               isSurfaceAvailableForRendering = aClass.getDeclaredField("isSurfaceAvailableForRendering");
+                            Class<? extends FlutterTextureView> aClass =
+                                    flutterTextureView.getClass();
+                            Field isSurfaceAvailableForRendering = aClass.getDeclaredField(
+                                    "isSurfaceAvailableForRendering");
                             isSurfaceAvailableForRendering.setAccessible(true);
                             isSurfaceAvailableForRendering.set(flutterTextureView, false);
-
-                            Field isAttachedToFlutterRenderer = aClass.getDeclaredField("isAttachedToFlutterRenderer");
-                            isAttachedToFlutterRenderer.setAccessible(true);
                         } catch (Exception e) {
                             // https://github.com/alibaba/flutter_boost/issues/1560
                             throw new RuntimeException("You *SHOULD* keep FlutterTextureView: -keep class io.flutter.embedding.android.FlutterTextureView { *; }.", e);