+++ example/android/app/src/main/java/io/flutter/embedding/android/LifecycleView.java
@@ -77,7 +77,7 @@ public class LifecycleView extends FrameLayout implements LifecycleOwner, Flutte
 
   public void onResume() {
     mLifecycleRegistry.handleLifecycleEvent(Lifecycle.Event.ON_RESUME);
-    platformPlugin = new PlatformPlugin(getActivity(), getFlutterEngine().getPlatformChannel());
+    platformPlugin = new PlatformPlugin(getActivity(), getFlutterEngine().getPlatformChannel(), this);
     mDelegate.onResume();
   }
 
@@ -174,11 +174,6 @@ public class LifecycleView extends FrameLayout implements LifecycleOwner, Flutte
   }
 
   @Nullable
-  public SplashScreen provideSplashScreen() {
-    return null;
-  }
-
-  @Nullable
   public FlutterEngine provideFlutterEngine(@NonNull Context context) {
     return null;
   }
@@ -197,6 +192,10 @@ public class LifecycleView extends FrameLayout implements LifecycleOwner, Flutte
 
   public boolean shouldAttachEngineToActivity() {
     return true;
+  }
+
+  public boolean attachToEngineAutomatically() {
+    return false;
   }
 
   public boolean shouldRestoreAndSaveState() {